django_app/requirements.txt
# Web Framework
Django==5.0.1
djangorestframework==3.14.0
uvicorn==0.29.0

# Database
mysql-connector-python==8.4.0

# Cache
redis==5.0.7

# Async Support
aiofiles==24.1.0
aiohttp==3.9.5

# Security & Authentication
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1

# Utils
shortuuid==1.0.13